Cross-sectional dependence in models

This routine computes all models considered for cross-sectional dependence using endogenously generated weight matrices. Currently "lm" for linear regression, "logit" (logit model), "poisson" (poisson model), "probit" (probit model), "negbin" (negative binomial) are supported. Note: the data should first be sorted by unit ID Pid, then by time ID Tid

ncd_gen(pars, Y, X, Xm = NULL, Xi = X, Tid, Pid, fun, k = 1, nt,
  utid, modclass = "lm", crit_obs = FALSE)

Arguments

pars
vector of parameter values at which to evaluate the function

Y
outcome variable

X
matrix of covariates or design matrix

Xm
matrix of other control variables

Xi
matrix of variable(s) that drive cross-sectional dependence

Tid
time ID

Pid
unit ID

fun
a function that parameterises the form of dependence. it can be user-written

k
order of polynomial approximation for fun()

nt
number of terms in approximation; should be of length equal to output of fun()

utid
unique time IDs to be used in the estimation; should leave at least 1 period for lagged Xi

modclass
the class of model. See description above for classes supported.

crit_obs
a logical for vector of observation-specific criterion values, squared residuals for "lm", negative log-likelihoods for MLE methods, etc.

Value

criterion value or vector of observation-specific criterion values if crit_obs=TRUE